% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\docType{package}
\name{classroom}
\alias{classroom}
\title{Google Classroom API
Manages classes, rosters, and invitations in Google Classroom.}
\details{
Authentication scopes used are:
\itemize{
\item https://www.googleapis.com/auth/classroom.push-notifications
\item https://www.googleapis.com/auth/classroom.rosters
\item https://www.googleapis.com/auth/classroom.topics
\item https://www.googleapis.com/auth/classroom.topics.readonly
\item https://www.googleapis.com/auth/classroom.student-submissions.students.readonly
\item https://www.googleapis.com/auth/classroom.guardianlinks.students.readonly
\item https://www.googleapis.com/auth/classroom.courses
\item https://www.googleapis.com/auth/classroom.courses.readonly
\item https://www.googleapis.com/auth/classroom.announcements
\item https://www.googleapis.com/auth/classroom.profile.photos
\item https://www.googleapis.com/auth/classroom.rosters.readonly
\item https://www.googleapis.com/auth/classroom.announcements.readonly
\item https://www.googleapis.com/auth/classroom.guardianlinks.students
\item https://www.googleapis.com/auth/classroom.student-submissions.me.readonly
\item https://www.googleapis.com/auth/classroom.coursework.students
\item https://www.googleapis.com/auth/classroom.coursework.students.readonly
\item https://www.googleapis.com/auth/classroom.guardianlinks.me.readonly
\item https://www.googleapis.com/auth/classroom.coursework.me.readonly
\item https://www.googleapis.com/auth/classroom.profile.emails
\item https://www.googleapis.com/auth/classroom.coursework.me
}
}
