% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleclassroom.R
\name{classroomService}
\alias{classroomService}
\title{Authenticate user and create classroom service}
\usage{
classroomService()
}
\value{

}
\description{
Authenticate user and create classroom service
}
\examples{
cs <- classroomService()
}
