% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Course}
\alias{Course}
\title{Course Object}
\usage{
Course(
  updateTime = NULL,
  calendarId = NULL,
  alternateLink = NULL,
  guardiansEnabled = NULL,
  ownerId = NULL,
  courseState = NULL,
  description = NULL,
  teacherGroupEmail = NULL,
  creationTime = NULL,
  teacherFolder = NULL,
  name = NULL,
  section = NULL,
  id = NULL,
  room = NULL,
  courseGroupEmail = NULL,
  enrollmentCode = NULL,
  courseMaterialSets = NULL,
  descriptionHeading = NULL
)
}
\arguments{
\item{updateTime}{Time of the most recent update to this course}

\item{calendarId}{The Calendar ID for a calendar that all course members can see, to which}

\item{alternateLink}{Absolute link to this course in the Classroom web UI}

\item{guardiansEnabled}{Whether or not guardian notifications are enabled for this course}

\item{ownerId}{The identifier of the owner of a course}

\item{courseState}{State of the course}

\item{description}{Optional description}

\item{teacherGroupEmail}{The email address of a Google group containing all teachers of the course}

\item{creationTime}{Creation time of the course}

\item{teacherFolder}{Information about a Drive Folder that is shared with all teachers of the}

\item{name}{Name of the course}

\item{section}{Section of the course}

\item{id}{Identifier for this course assigned by Classroom}

\item{room}{Optional room location}

\item{courseGroupEmail}{The email address of a Google group containing all members of the course}

\item{enrollmentCode}{Enrollment code to use when joining this course}

\item{courseMaterialSets}{Sets of materials that appear on the 'about' page of this course}

\item{descriptionHeading}{Optional heading for the description}
}
\value{
Course object
}
\description{
Course Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Course in Classroom.
}
\seealso{
Other Course functions: 
\code{\link{courses.create}()},
\code{\link{courses.patch}()},
\code{\link{courses.update}()}
}
\concept{Course functions}
