% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{CourseAlias}
\alias{CourseAlias}
\title{CourseAlias Object}
\usage{
CourseAlias(alias = NULL)
}
\arguments{
\item{alias}{Alias string}
}
\value{
CourseAlias object
}
\description{
CourseAlias Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Alternative identifier for a course.An alias uniquely identifies a course. It must be unique within one of thefollowing scopes:* domain: A domain-scoped alias is visible to all users within the aliascreator's domain and can be created only by a domain admin. A domain-scopedalias is often used when a course has an identifier external to Classroom.* project: A project-scoped alias is visible to any request from anapplication using the Developer Console project ID that created the aliasand can be created by any project. A project-scoped alias is often used whenan application has alternative identifiers. A random value can also be usedto avoid duplicate courses in the event of transmission failures, as retryinga request will return \code{ALREADY_EXISTS} if a previous one has succeeded.
}
\seealso{
Other CourseAlias functions: 
\code{\link{courses.aliases.create}()}
}
\concept{CourseAlias functions}
