% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{CourseMaterial}
\alias{CourseMaterial}
\title{CourseMaterial Object}
\usage{
CourseMaterial(link = NULL, driveFile = NULL, youTubeVideo = NULL, form = NULL)
}
\arguments{
\item{link}{Link atatchment}

\item{driveFile}{Google Drive file attachment}

\item{youTubeVideo}{Youtube video attachment}

\item{form}{Google Forms attachment}
}
\value{
CourseMaterial object
}
\description{
CourseMaterial Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A material attached to a course as part of a material set.
}
\concept{CourseMaterial functions}
