% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{CourseRosterChangesInfo}
\alias{CourseRosterChangesInfo}
\title{CourseRosterChangesInfo Object}
\usage{
CourseRosterChangesInfo(courseId = NULL)
}
\arguments{
\item{courseId}{The \code{course_id} of the course to subscribe to roster changes for}
}
\value{
CourseRosterChangesInfo object
}
\description{
CourseRosterChangesInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a \code{Feed} with a \code{feed_type} of \code{COURSE_ROSTER_CHANGES}.
}
\concept{CourseRosterChangesInfo functions}
