% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.aliases.create}
\alias{courses.aliases.create}
\title{Creates an alias for a course.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to create thealias or for access errors.* \code{NOT_FOUND} if the course does not exist.* \code{ALREADY_EXISTS} if the alias already exists.* \code{FAILED_PRECONDITION} if the alias requested does not make sense for the  requesting user or course (for example, if a user not in a domain  attempts to access a domain-scoped alias).}
\usage{
courses.aliases.create(CourseAlias, courseId)
}
\arguments{
\item{CourseAlias}{The \link{CourseAlias} object to pass to this method}

\item{courseId}{Identifier of the course to alias}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.courses
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.courses)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other CourseAlias functions: 
\code{\link{CourseAlias}()}
}
\concept{CourseAlias functions}
