% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.aliases.delete}
\alias{courses.aliases.delete}
\title{Deletes an alias of a course.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to remove thealias or for access errors.* \code{NOT_FOUND} if the alias does not exist.* \code{FAILED_PRECONDITION} if the alias requested does not make sense for the  requesting user or course (for example, if a user not in a domain  attempts to delete a domain-scoped alias).}
\usage{
courses.aliases.delete(courseId, alias)
}
\arguments{
\item{courseId}{Identifier of the course whose alias should be deleted}

\item{alias}{Alias to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.courses
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.courses)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
