% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.aliases.list}
\alias{courses.aliases.list}
\title{Returns a list of aliases for a course.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to access thecourse or for access errors.* \code{NOT_FOUND} if the course does not exist.}
\usage{
courses.aliases.list(courseId, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{courseId}{The identifier of the course}

\item{pageToken}{nextPageToken}

\item{pageSize}{Maximum number of items to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.courses
\item https://www.googleapis.com/auth/classroom.courses.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.courses, https://www.googleapis.com/auth/classroom.courses.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
