% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.announcements.create}
\alias{courses.announcements.create}
\title{Creates an announcement.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course, create announcements in the requested course, share aDrive attachment, or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course does not exist.* \code{FAILED_PRECONDITION} for the following request error:    * AttachmentNotVisible}
\usage{
courses.announcements.create(Announcement, courseId)
}
\arguments{
\item{Announcement}{The \link{Announcement} object to pass to this method}

\item{courseId}{Identifier of the course}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.announcements
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.announcements)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Announcement functions: 
\code{\link{Announcement}()},
\code{\link{courses.announcements.patch}()}
}
\concept{Announcement functions}
