% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.announcements.delete}
\alias{courses.announcements.delete}
\title{Deletes an announcement.This request must be made by the Developer Console project of the\href{https://support.google.com/cloud/answer/6158849}{OAuth client ID} used tocreate the corresponding announcement item.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting developer project did not createthe corresponding announcement, if the requesting user is not permittedto delete the requested course or for access errors.* \code{FAILED_PRECONDITION} if the requested announcement has already beendeleted.* \code{NOT_FOUND} if no course exists with the requested ID.}
\usage{
courses.announcements.delete(courseId, id)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{id}{Identifier of the announcement to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.announcements
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.announcements)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
