% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.announcements.list}
\alias{courses.announcements.list}
\title{Returns a list of announcements that the requester is permitted to view.Course students may only view \code{PUBLISHED} announcements. Course teachersand domain administrators may view all announcements.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to accessthe requested course or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course does not exist.}
\usage{
courses.announcements.list(
  courseId,
  announcementStates = NULL,
  orderBy = NULL,
  pageToken = NULL,
  pageSize = NULL
)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{announcementStates}{Restriction on the \code{state} of announcements returned}

\item{orderBy}{Optional sort ordering for results}

\item{pageToken}{nextPageToken}

\item{pageSize}{Maximum number of items to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.announcements
\item https://www.googleapis.com/auth/classroom.announcements.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.announcements, https://www.googleapis.com/auth/classroom.announcements.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
