% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.announcements.modifyAssignees}
\alias{courses.announcements.modifyAssignees}
\title{Modifies assignee mode and options of an announcement.Only a teacher of the course that contains the announcement maycall this method.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course or course work or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course or course work does not exist.}
\usage{
courses.announcements.modifyAssignees(
  ModifyAnnouncementAssigneesRequest,
  courseId,
  id
)
}
\arguments{
\item{ModifyAnnouncementAssigneesRequest}{The \link{ModifyAnnouncementAssigneesRequest} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{id}{Identifier of the announcement}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.announcements
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.announcements)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other ModifyAnnouncementAssigneesRequest functions: 
\code{\link{ModifyAnnouncementAssigneesRequest}()}
}
\concept{ModifyAnnouncementAssigneesRequest functions}
