% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.announcements.patch}
\alias{courses.announcements.patch}
\title{Updates one or more fields of an announcement.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting developer project did not createthe corresponding announcement or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{FAILED_PRECONDITION} if the requested announcement has already beendeleted.* \code{NOT_FOUND} if the requested course or announcement does not exist}
\usage{
courses.announcements.patch(Announcement, courseId, id, updateMask = NULL)
}
\arguments{
\item{Announcement}{The \link{Announcement} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{id}{Identifier of the announcement}

\item{updateMask}{Mask that identifies which fields on the announcement to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.announcements
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.announcements)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Announcement functions: 
\code{\link{Announcement}()},
\code{\link{courses.announcements.create}()}
}
\concept{Announcement functions}
