% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.courseWork.create}
\alias{courses.courseWork.create}
\title{Creates course work.The resulting course work (and corresponding student submissions) areassociated with the Developer Console project of the\href{https://support.google.com/cloud/answer/6158849}{OAuth client ID} used tomake the request. Classroom API requests to modify course work and studentsubmissions must be made with an OAuth client ID from the associatedDeveloper Console project.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course, create course work in the requested course, share aDrive attachment, or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course does not exist.* \code{FAILED_PRECONDITION} for the following request error:    * AttachmentNotVisible}
\usage{
courses.courseWork.create(CourseWork, courseId)
}
\arguments{
\item{CourseWork}{The \link{CourseWork} object to pass to this method}

\item{courseId}{Identifier of the course}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.students
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.coursework.students)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other CourseWork functions: 
\code{\link{CourseWork}()},
\code{\link{courses.courseWork.patch}()}
}
\concept{CourseWork functions}
