% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.courseWork.get}
\alias{courses.courseWork.get}
\title{Returns course work.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course or course work, or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course or course work does not exist.}
\usage{
courses.courseWork.get(courseId, id)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{id}{Identifier of the course work}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.me
\item https://www.googleapis.com/auth/classroom.coursework.me.readonly
\item https://www.googleapis.com/auth/classroom.coursework.students
\item https://www.googleapis.com/auth/classroom.coursework.students.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.coursework.me, https://www.googleapis.com/auth/classroom.coursework.me.readonly, https://www.googleapis.com/auth/classroom.coursework.students, https://www.googleapis.com/auth/classroom.coursework.students.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
