% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.courseWork.list}
\alias{courses.courseWork.list}
\title{Returns a list of course work that the requester is permitted to view.Course students may only view \code{PUBLISHED} course work. Course teachersand domain administrators may view all course work.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to accessthe requested course or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course does not exist.}
\usage{
courses.courseWork.list(
  courseId,
  orderBy = NULL,
  pageToken = NULL,
  pageSize = NULL,
  courseWorkStates = NULL
)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{orderBy}{Optional sort ordering for results}

\item{pageToken}{nextPageToken}

\item{pageSize}{Maximum number of items to return}

\item{courseWorkStates}{Restriction on the work status to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.me
\item https://www.googleapis.com/auth/classroom.coursework.me.readonly
\item https://www.googleapis.com/auth/classroom.coursework.students
\item https://www.googleapis.com/auth/classroom.coursework.students.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.coursework.me, https://www.googleapis.com/auth/classroom.coursework.me.readonly, https://www.googleapis.com/auth/classroom.coursework.students, https://www.googleapis.com/auth/classroom.coursework.students.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
