% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.courseWork.patch}
\alias{courses.courseWork.patch}
\title{Updates one or more fields of a course work.See google.classroom.v1.CourseWork for detailsof which fields may be updated and who may change them.This request must be made by the Developer Console project of the\href{https://support.google.com/cloud/answer/6158849}{OAuth client ID} used tocreate the corresponding course work item.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting developer project did not createthe corresponding course work, if the user is not permitted to make therequested modification to the student submission, or foraccess errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{FAILED_PRECONDITION} if the requested course work has already beendeleted.* \code{NOT_FOUND} if the requested course, course work, or student submissiondoes not exist.}
\usage{
courses.courseWork.patch(CourseWork, courseId, id, updateMask = NULL)
}
\arguments{
\item{CourseWork}{The \link{CourseWork} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{id}{Identifier of the course work}

\item{updateMask}{Mask that identifies which fields on the course work to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.students
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.coursework.students)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other CourseWork functions: 
\code{\link{CourseWork}()},
\code{\link{courses.courseWork.create}()}
}
\concept{CourseWork functions}
