% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.courseWork.studentSubmissions.list}
\alias{courses.courseWork.studentSubmissions.list}
\title{Returns a list of student submissions that the requester is permitted toview, factoring in the OAuth scopes of the request.\code{-} may be specified as the \code{course_work_id} to include studentsubmissions for multiple course work items.Course students may only view their own work. Course teachersand domain administrators may view all student submissions.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course or course work, or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course does not exist.}
\usage{
courses.courseWork.studentSubmissions.list(
  courseId,
  courseWorkId,
  userId = NULL,
  late = NULL,
  pageToken = NULL,
  states = NULL,
  pageSize = NULL
)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{courseWorkId}{Identifier of the student work to request}

\item{userId}{Optional argument to restrict returned student work to those owned by the}

\item{late}{Requested lateness value}

\item{pageToken}{nextPageToken}

\item{states}{Requested submission states}

\item{pageSize}{Maximum number of items to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.me
\item https://www.googleapis.com/auth/classroom.coursework.me.readonly
\item https://www.googleapis.com/auth/classroom.coursework.students
\item https://www.googleapis.com/auth/classroom.coursework.students.readonly
\item https://www.googleapis.com/auth/classroom.student-submissions.me.readonly
\item https://www.googleapis.com/auth/classroom.student-submissions.students.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.coursework.me, https://www.googleapis.com/auth/classroom.coursework.me.readonly, https://www.googleapis.com/auth/classroom.coursework.students, https://www.googleapis.com/auth/classroom.coursework.students.readonly, https://www.googleapis.com/auth/classroom.student-submissions.me.readonly, https://www.googleapis.com/auth/classroom.student-submissions.students.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
