% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.courseWork.studentSubmissions.patch}
\alias{courses.courseWork.studentSubmissions.patch}
\title{Updates one or more fields of a student submission.See google.classroom.v1.StudentSubmission for detailsof which fields may be updated and who may change them.This request must be made by the Developer Console project of the\href{https://support.google.com/cloud/answer/6158849}{OAuth client ID} used tocreate the corresponding course work item.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting developer project did not createthe corresponding course work, if the user is not permitted to make therequested modification to the student submission, or foraccess errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course, course work, or student submissiondoes not exist.}
\usage{
courses.courseWork.studentSubmissions.patch(
  StudentSubmission,
  courseId,
  courseWorkId,
  id,
  updateMask = NULL
)
}
\arguments{
\item{StudentSubmission}{The \link{StudentSubmission} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{courseWorkId}{Identifier of the course work}

\item{id}{Identifier of the student submission}

\item{updateMask}{Mask that identifies which fields on the student submission to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.me
\item https://www.googleapis.com/auth/classroom.coursework.students
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.coursework.me, https://www.googleapis.com/auth/classroom.coursework.students)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other StudentSubmission functions: 
\code{\link{StudentSubmission}()}
}
\concept{StudentSubmission functions}
