% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.courseWork.studentSubmissions.reclaim}
\alias{courses.courseWork.studentSubmissions.reclaim}
\title{Reclaims a student submission on behalf of the student that owns it.Reclaiming a student submission transfers ownership of attached Drivefiles to the student and updates the submission state.Only the student that owns the requested student submission may call thismethod, and only for a student submission that has been turned in.This request must be made by the Developer Console project of the\href{https://support.google.com/cloud/answer/6158849}{OAuth client ID} used tocreate the corresponding course work item.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course or course work, unsubmit the requested student submission,or for access errors.* \code{FAILED_PRECONDITION} if the student submission has not been turned in.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course, course work, or student submissiondoes not exist.}
\usage{
courses.courseWork.studentSubmissions.reclaim(
  ReclaimStudentSubmissionRequest,
  courseId,
  courseWorkId,
  id
)
}
\arguments{
\item{ReclaimStudentSubmissionRequest}{The \link{ReclaimStudentSubmissionRequest} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{courseWorkId}{Identifier of the course work}

\item{id}{Identifier of the student submission}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.me
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.coursework.me)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other ReclaimStudentSubmissionRequest functions: 
\code{\link{ReclaimStudentSubmissionRequest}()}
}
\concept{ReclaimStudentSubmissionRequest functions}
