% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.courseWork.studentSubmissions.turnIn}
\alias{courses.courseWork.studentSubmissions.turnIn}
\title{Turns in a student submission.Turning in a student submission transfers ownership of attached Drivefiles to the teacher and may also update the submission state.This may only be called by the student that owns the specified studentsubmission.This request must be made by the Developer Console project of the\href{https://support.google.com/cloud/answer/6158849}{OAuth client ID} used tocreate the corresponding course work item.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course or course work, turn in the requested student submission,or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course, course work, or student submissiondoes not exist.}
\usage{
courses.courseWork.studentSubmissions.turnIn(
  TurnInStudentSubmissionRequest,
  courseId,
  courseWorkId,
  id
)
}
\arguments{
\item{TurnInStudentSubmissionRequest}{The \link{TurnInStudentSubmissionRequest} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{courseWorkId}{Identifier of the course work}

\item{id}{Identifier of the student submission}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.me
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.coursework.me)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other TurnInStudentSubmissionRequest functions: 
\code{\link{TurnInStudentSubmissionRequest}()}
}
\concept{TurnInStudentSubmissionRequest functions}
