% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.create}
\alias{courses.create}
\title{Creates a course.The user specified in \code{ownerId} is the owner of the created courseand added as a teacher.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to createcourses or for access errors.* \code{NOT_FOUND} if the primary teacher is not a valid user.* \code{FAILED_PRECONDITION} if the course owner's account is disabled or forthe following request errors:    * UserGroupsMembershipLimitReached* \code{ALREADY_EXISTS} if an alias was specified in the \code{id} andalready exists.}
\usage{
courses.create(Course)
}
\arguments{
\item{Course}{The \link{Course} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.courses
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.courses)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Course functions: 
\code{\link{Course}()},
\code{\link{courses.patch}()},
\code{\link{courses.update}()}
}
\concept{Course functions}
