% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.list}
\alias{courses.list}
\title{Returns a list of courses that the requesting user is permitted to view,restricted to those that match the request. Returned courses are ordered bycreation time, with the most recently created coming first.This method returns the following error codes:* \code{PERMISSION_DENIED} for access errors.* \code{INVALID_ARGUMENT} if the query argument is malformed.* \code{NOT_FOUND} if any users specified in the query arguments do not exist.}
\usage{
courses.list(
  studentId = NULL,
  pageToken = NULL,
  pageSize = NULL,
  teacherId = NULL,
  courseStates = NULL
)
}
\arguments{
\item{studentId}{Restricts returned courses to those having a student with the specified}

\item{pageToken}{nextPageToken}

\item{pageSize}{Maximum number of items to return}

\item{teacherId}{Restricts returned courses to those having a teacher with the specified}

\item{courseStates}{Restricts returned courses to those in one of the specified states}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.courses
\item https://www.googleapis.com/auth/classroom.courses.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.courses, https://www.googleapis.com/auth/classroom.courses.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
