% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.patch}
\alias{courses.patch}
\title{Updates one or more fields in a course.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to modify therequested course or for access errors.* \code{NOT_FOUND} if no course exists with the requested ID.* \code{INVALID_ARGUMENT} if invalid fields are specified in the update mask orif no update mask is supplied.* \code{FAILED_PRECONDITION} for the following request errors:    * CourseNotModifiable}
\usage{
courses.patch(Course, id, updateMask = NULL)
}
\arguments{
\item{Course}{The \link{Course} object to pass to this method}

\item{id}{Identifier of the course to update}

\item{updateMask}{Mask that identifies which fields on the course to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.courses
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.courses)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Course functions: 
\code{\link{Course}()},
\code{\link{courses.create}()},
\code{\link{courses.update}()}
}
\concept{Course functions}
