% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.students.create}
\alias{courses.students.create}
\title{Adds a user as a student of a course.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to createstudents in this course or for access errors.* \code{NOT_FOUND} if the requested course ID does not exist.* \code{FAILED_PRECONDITION} if the requested user's account is disabled,for the following request errors:    * CourseMemberLimitReached    * CourseNotModifiable    * UserGroupsMembershipLimitReached* \code{ALREADY_EXISTS} if the user is already a student or teacher in thecourse.}
\usage{
courses.students.create(Student, courseId, enrollmentCode = NULL)
}
\arguments{
\item{Student}{The \link{Student} object to pass to this method}

\item{courseId}{Identifier of the course to create the student in}

\item{enrollmentCode}{Enrollment code of the course to create the student in}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.profile.emails
\item https://www.googleapis.com/auth/classroom.profile.photos
\item https://www.googleapis.com/auth/classroom.rosters
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.profile.emails, https://www.googleapis.com/auth/classroom.profile.photos, https://www.googleapis.com/auth/classroom.rosters)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Student functions: 
\code{\link{Student}()}
}
\concept{Student functions}
