% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.students.list}
\alias{courses.students.list}
\title{Returns a list of students of this course that the requesteris permitted to view.This method returns the following error codes:* \code{NOT_FOUND} if the course does not exist.* \code{PERMISSION_DENIED} for access errors.}
\usage{
courses.students.list(courseId, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{pageToken}{nextPageToken}

\item{pageSize}{Maximum number of items to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.profile.emails
\item https://www.googleapis.com/auth/classroom.profile.photos
\item https://www.googleapis.com/auth/classroom.rosters
\item https://www.googleapis.com/auth/classroom.rosters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.profile.emails, https://www.googleapis.com/auth/classroom.profile.photos, https://www.googleapis.com/auth/classroom.rosters, https://www.googleapis.com/auth/classroom.rosters.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
