% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.teachers.create}
\alias{courses.teachers.create}
\title{Creates a teacher of a course.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not  permitted to createteachers in this course or for access errors.* \code{NOT_FOUND} if the requested course ID does not exist.* \code{FAILED_PRECONDITION} if the requested user's account is disabled,for the following request errors:    * CourseMemberLimitReached    * CourseNotModifiable    * CourseTeacherLimitReached    * UserGroupsMembershipLimitReached* \code{ALREADY_EXISTS} if the user is already a teacher or student in thecourse.}
\usage{
courses.teachers.create(Teacher, courseId)
}
\arguments{
\item{Teacher}{The \link{Teacher} object to pass to this method}

\item{courseId}{Identifier of the course}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.profile.emails
\item https://www.googleapis.com/auth/classroom.profile.photos
\item https://www.googleapis.com/auth/classroom.rosters
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.profile.emails, https://www.googleapis.com/auth/classroom.profile.photos, https://www.googleapis.com/auth/classroom.rosters)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Teacher functions: 
\code{\link{Teacher}()}
}
\concept{Teacher functions}
