% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.teachers.get}
\alias{courses.teachers.get}
\title{Returns a teacher of a course.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to viewteachers of this course or for access errors.* \code{NOT_FOUND} if no teacher of this course has the requested ID or if thecourse does not exist.}
\usage{
courses.teachers.get(courseId, userId)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{userId}{Identifier of the teacher to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.profile.emails
\item https://www.googleapis.com/auth/classroom.profile.photos
\item https://www.googleapis.com/auth/classroom.rosters
\item https://www.googleapis.com/auth/classroom.rosters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.profile.emails, https://www.googleapis.com/auth/classroom.profile.photos, https://www.googleapis.com/auth/classroom.rosters, https://www.googleapis.com/auth/classroom.rosters.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
