% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.topics.create}
\alias{courses.topics.create}
\title{Creates a topic.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course, create a topic in the requested course,or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course does not exist.}
\usage{
courses.topics.create(Topic, courseId)
}
\arguments{
\item{Topic}{The \link{Topic} object to pass to this method}

\item{courseId}{Identifier of the course}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.topics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.topics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Topic functions: 
\code{\link{Topic}()},
\code{\link{courses.topics.patch}()}
}
\concept{Topic functions}
