% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.topics.delete}
\alias{courses.topics.delete}
\title{Deletes a topic.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not allowed to delete therequested topic or for access errors.* \code{FAILED_PRECONDITION} if the requested topic has already beendeleted.* \code{NOT_FOUND} if no course or topic exists with the requested ID.}
\usage{
courses.topics.delete(courseId, id)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{id}{Identifier of the topic to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.topics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.topics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
