% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.topics.get}
\alias{courses.topics.get}
\title{Returns a topic.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course or topic, or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course or topic does not exist.}
\usage{
courses.topics.get(courseId, id)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{id}{Identifier of the topic}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.topics
\item https://www.googleapis.com/auth/classroom.topics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.topics, https://www.googleapis.com/auth/classroom.topics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
