% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{courses.topics.patch}
\alias{courses.topics.patch}
\title{Updates one or more fields of a topic.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting developer project did not createthe corresponding topic or for access errors.* \code{INVALID_ARGUMENT} if the request is malformed.* \code{NOT_FOUND} if the requested course or topic does not exist}
\usage{
courses.topics.patch(Topic, courseId, id, updateMask = NULL)
}
\arguments{
\item{Topic}{The \link{Topic} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{id}{Identifier of the topic}

\item{updateMask}{Mask that identifies which fields on the topic to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.topics
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.topics)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Topic functions: 
\code{\link{Topic}()},
\code{\link{courses.topics.create}()}
}
\concept{Topic functions}
