% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{CourseWork}
\alias{CourseWork}
\title{CourseWork Object}
\usage{
CourseWork(
  materials = NULL,
  associatedWithDeveloper = NULL,
  updateTime = NULL,
  alternateLink = NULL,
  assigneeMode = NULL,
  maxPoints = NULL,
  workType = NULL,
  assignment = NULL,
  multipleChoiceQuestion = NULL,
  description = NULL,
  scheduledTime = NULL,
  creationTime = NULL,
  creatorUserId = NULL,
  individualStudentsOptions = NULL,
  dueDate = NULL,
  submissionModificationMode = NULL,
  state = NULL,
  courseId = NULL,
  id = NULL,
  dueTime = NULL,
  title = NULL,
  topicId = NULL
)
}
\arguments{
\item{materials}{Additional materials}

\item{associatedWithDeveloper}{Whether this course work item is associated with the Developer Console}

\item{updateTime}{Timestamp of the most recent change to this course work}

\item{alternateLink}{Absolute link to this course work in the Classroom web UI}

\item{assigneeMode}{Assignee mode of the coursework}

\item{maxPoints}{Maximum grade for this course work}

\item{workType}{Type of this course work}

\item{assignment}{Assignment details}

\item{multipleChoiceQuestion}{Multiple choice question details}

\item{description}{Optional description of this course work}

\item{scheduledTime}{Optional timestamp when this course work is scheduled to be published}

\item{creationTime}{Timestamp when this course work was created}

\item{creatorUserId}{Identifier for the user that created the coursework}

\item{individualStudentsOptions}{Identifiers of students with access to the coursework}

\item{dueDate}{Optional date, in UTC, that submissions for this course work are due}

\item{submissionModificationMode}{Setting to determine when students are allowed to modify submissions}

\item{state}{Status of this course work}

\item{courseId}{Identifier of the course}

\item{id}{Classroom-assigned identifier of this course work, unique per course}

\item{dueTime}{Optional time of day, in UTC, that submissions for this course work}

\item{title}{Title of this course work}

\item{topicId}{Identifier for the topic that this coursework is associated with}
}
\value{
CourseWork object
}
\description{
CourseWork Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Course work created by a teacher for students of the course.
}
\seealso{
Other CourseWork functions: 
\code{\link{courses.courseWork.create}()},
\code{\link{courses.courseWork.patch}()}
}
\concept{CourseWork functions}
