% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{CourseWorkChangesInfo}
\alias{CourseWorkChangesInfo}
\title{CourseWorkChangesInfo Object}
\usage{
CourseWorkChangesInfo(courseId = NULL)
}
\arguments{
\item{courseId}{The \code{course_id} of the course to subscribe to work changes for}
}
\value{
CourseWorkChangesInfo object
}
\description{
CourseWorkChangesInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a \code{Feed} with a \code{feed_type} of \code{COURSE_WORK_CHANGES}.
}
\concept{CourseWorkChangesInfo functions}
