% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Date}
\alias{Date}
\title{Date Object}
\usage{
Date(year = NULL, day = NULL, month = NULL)
}
\arguments{
\item{year}{Year of date}

\item{day}{Day of month}

\item{month}{Month of year}
}
\value{
Date object
}
\description{
Date Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a whole or partial calendar date, e.g. a birthday. The time of dayand time zone are either specified elsewhere or are not significant. The dateis relative to the Proleptic Gregorian Calendar. This can represent:* A full date, with non-zero year, month and day values* A month and day value, with a zero year, e.g. an anniversary* A year on its own, with zero month and day values* A year and month value, with a zero day, e.g. a credit card expiration dateRelated types are google.type.TimeOfDay and \code{google.protobuf.Timestamp}.
}
\concept{Date functions}
