% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursework.R
\name{download_studentRmds}
\alias{download_studentRmds}
\title{Download student Rmd submissions}
\usage{
download_studentRmds(title, df_studentSubmissions)
}
\arguments{
\item{title}{A character. The title of the course work. It will be used as a folder name for this course work}

\item{df_studentSubmissions}{A data frame. Can be the return from get_studentSubmissionForACourseWork}
}
\value{

}
\description{
Download student Rmd submissions
}
\examples{
none
}
