% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{DriveFolder}
\alias{DriveFolder}
\title{DriveFolder Object}
\usage{
DriveFolder(title = NULL, alternateLink = NULL, id = NULL)
}
\arguments{
\item{title}{Title of the Drive folder}

\item{alternateLink}{URL that can be used to access the Drive folder}

\item{id}{Drive API resource ID}
}
\value{
DriveFolder object
}
\description{
DriveFolder Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Representation of a Google Drive folder.
}
\concept{DriveFolder functions}
