% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_announcement}
\alias{gc_announcement}
\title{Announcement Object}
\usage{
gc_announcement(
  scheduledTime = NULL,
  creationTime = NULL,
  creatorUserId = NULL,
  individualStudentsOptions = NULL,
  state = NULL,
  text = NULL,
  courseId = NULL,
  id = NULL,
  materials = NULL,
  updateTime = NULL,
  assigneeMode = NULL,
  alternateLink = NULL
)
}
\arguments{
\item{scheduledTime}{Optional timestamp when this announcement is scheduled to be published}

\item{creationTime}{Timestamp when this announcement was created}

\item{creatorUserId}{Identifier for the user that created the announcement}

\item{individualStudentsOptions}{Identifiers of students with access to the announcement}

\item{state}{Status of this announcement}

\item{text}{Description of this announcement}

\item{courseId}{Identifier of the course}

\item{id}{Classroom-assigned identifier of this announcement, unique per course}

\item{materials}{Additional materials}

\item{updateTime}{Timestamp of the most recent change to this announcement}

\item{assigneeMode}{Assignee mode of the announcement}

\item{alternateLink}{Absolute link to this announcement in the Classroom web UI}
}
\value{
Announcement object
}
\description{
Announcement Object
}
\details{
Announcement created by a teacher for students of the course
}
\seealso{
Other Announcement functions: 
\code{\link{gc_courses_announcements_create}()},
\code{\link{gc_courses_announcements_patch}()}
}
\concept{Announcement functions}
