% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_assignment}
\alias{gc_assignment}
\title{Assignment Object}
\usage{
gc_assignment(studentWorkFolder = NULL)
}
\arguments{
\item{studentWorkFolder}{Drive folder where attachments from student submissions are placed}
}
\value{
Assignment object
}
\description{
Assignment Object
}
\details{
Additional details for assignments.
}
\concept{Assignment functions}
