% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_assignment_submission}
\alias{gc_assignment_submission}
\title{AssignmentSubmission Object}
\usage{
gc_assignment_submission(attachments = NULL)
}
\arguments{
\item{attachments}{Attachments added by the student}
}
\value{
AssignmentSubmission object
}
\description{
AssignmentSubmission Object
}
\details{
Student work for an assignment.
}
\concept{AssignmentSubmission functions}
