% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_cloud_pubsub_topic}
\alias{gc_cloud_pubsub_topic}
\title{CloudPubsubTopic Object}
\usage{
gc_cloud_pubsub_topic(topicName = NULL)
}
\arguments{
\item{topicName}{The \code{name} field of a Cloud Pub/Sub}
}
\value{
CloudPubsubTopic object
}
\description{
CloudPubsubTopic Object
}
\details{
A reference to a Cloud Pub/Sub topic.To register for notifications, the owner of the topic must grant\code{classroom-notifications@system.gserviceaccount.com} the \code{projects.topics.publish} permission.
}
\concept{CloudPubsubTopic functions}
