% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_course}
\alias{gc_course}
\title{Course Object}
\usage{
gc_course(
  id = NULL,
  name = NULL,
  section = NULL,
  descriptionHeading = NULL,
  ownerId = NULL,
  creationTime = NULL,
  updateTime = NULL,
  enrollmentCode = NULL,
  courseState = NULL,
  alternateLink = NULL,
  teacherGroupEmail = NULL,
  courseGroupEmail = NULL,
  teacherFolder = NULL,
  guardiansEnabled = NULL,
  calendarId = NULL,
  room = NULL,
  description = NULL,
  courseMaterialSets = NULL
)
}
\arguments{
\item{id}{Identifier for this course assigned by Classroom}

\item{name}{Name of the course}

\item{section}{Section of the course}

\item{descriptionHeading}{Optional heading for the description}

\item{ownerId}{The identifier of the owner of a course}

\item{creationTime}{Creation time of the course}

\item{updateTime}{Time of the most recent update to this course}

\item{enrollmentCode}{Enrollment code to use when joining this course}

\item{courseState}{State of the course}

\item{alternateLink}{Absolute link to this course in the Classroom web UI}

\item{teacherGroupEmail}{The email address of a Google group containing all teachers of the course}

\item{courseGroupEmail}{The email address of a Google group containing all members of the course}

\item{teacherFolder}{Information about a Drive Folder that is shared with all teachers of the}

\item{guardiansEnabled}{Whether or not guardian notifications are enabled for this course}

\item{calendarId}{The Calendar ID for a calendar that all course members can see, to which}

\item{room}{Optional room location}

\item{description}{Optional description}

\item{courseMaterialSets}{Sets of materials that appear on the 'about' page of this course}
}
\value{
Course object
}
\description{
See \href{https://developers.google.com/classroom/reference/rest/v1/courses#resource:-course}{Google Documentation} for details
}
\details{
A Course in Classroom.
}
\seealso{
Other Course functions: 
\code{\link{gc_courses_create}()},
\code{\link{gc_courses_patch}()},
\code{\link{gc_courses_update}()}
}
\concept{Course functions}
