% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_course_material}
\alias{gc_course_material}
\title{CourseMaterial Object}
\usage{
gc_course_material(
  link = NULL,
  driveFile = NULL,
  youTubeVideo = NULL,
  form = NULL
)
}
\arguments{
\item{link}{Link atatchment}

\item{driveFile}{Google Drive file attachment}

\item{youTubeVideo}{Youtube video attachment}

\item{form}{Google Forms attachment}
}
\value{
CourseMaterial object
}
\description{
CourseMaterial Object
}
\details{
A material attached to a course as part of a material set.
}
\concept{CourseMaterial functions}
