% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_course_material_set}
\alias{gc_course_material_set}
\title{CourseMaterialSet Object}
\usage{
gc_course_material_set(title = NULL, materials = NULL)
}
\arguments{
\item{title}{Title for this set}

\item{materials}{Materials attached to this set}
}
\value{
CourseMaterialSet object
}
\description{
CourseMaterialSet Object
}
\details{
A set of materials that appears on the 'About' page of the course.These materials might include a syllabus, schedule, or other backgroundinformation relating to the course as a whole.
}
\concept{CourseMaterialSet functions}
