% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_course_roster_changes_info}
\alias{gc_course_roster_changes_info}
\title{CourseRosterChangesInfo Object}
\usage{
gc_course_roster_changes_info(courseId = NULL)
}
\arguments{
\item{courseId}{The \code{course_id} of the course to subscribe to roster changes for}
}
\value{
CourseRosterChangesInfo object
}
\description{
CourseRosterChangesInfo Object
}
\details{
Information about a \code{Feed} with a \code{feed_type} of \code{COURSE_ROSTER_CHANGES}.
}
\concept{CourseRosterChangesInfo functions}
