% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_course_work_changes_info}
\alias{gc_course_work_changes_info}
\title{CourseWorkChangesInfo Object}
\usage{
gc_course_work_changes_info(courseId = NULL)
}
\arguments{
\item{courseId}{The \code{course_id} of the course to subscribe to work changes for}
}
\value{
CourseWorkChangesInfo object
}
\description{
CourseWorkChangesInfo Object
}
\details{
Information about a \code{Feed} with a \code{feed_type} of \code{COURSE_WORK_CHANGES}.
}
\concept{CourseWorkChangesInfo functions}
