% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_aliases_list}
\alias{gc_courses_aliases_list}
\title{Aliases List}
\usage{
gc_courses_aliases_list(courseId, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{courseId}{The identifier of the course}

\item{pageToken}{nextPageToken}

\item{pageSize}{Maximum number of items to return}
}
\description{
Returns a list of aliases for a course. This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to access thecourse or for access errors.
\item \code{NOT_FOUND} if the course does not exist.
}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.courses
\item https://www.googleapis.com/auth/classroom.courses.readonly
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.courses", "https://www.googleapis.com/auth/classroom.courses.readonly")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
