% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_announcements_create}
\alias{gc_courses_announcements_create}
\title{Creates an announcement}
\usage{
gc_courses_announcements_create(Announcement, courseId)
}
\arguments{
\item{Announcement}{The \link{gc_announcement} object to pass to this method}

\item{courseId}{Identifier of the course}
}
\description{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course, create announcements in the requested course, share aDrive attachment, or for access errors.
\item \code{INVALID_ARGUMENT} if the request is malformed.
\item \code{NOT_FOUND} if the requested course does not exist.
\item \code{FAILED_PRECONDITION} for the following request error:
\itemize{
\item AttachmentNotVisible
}
}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.announcements
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.announcements")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Announcement functions: 
\code{\link{gc_announcement}()},
\code{\link{gc_courses_announcements_patch}()}
}
\concept{Announcement functions}
