% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_announcements_delete}
\alias{gc_courses_announcements_delete}
\title{Deletes an announcement}
\usage{
gc_courses_announcements_delete(courseId, id)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{id}{Identifier of the announcement to delete}
}
\description{
This request must be made by the Developer Console project of the \href{https://support.google.com/cloud/answer/6158849}{OAuth client ID} used to
create the corresponding announcement item.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting developer project did not createthe corresponding announcement, if the requesting user is not permitted to delete the requested course or for access errors.
\item \code{FAILED_PRECONDITION} if the requested announcement has already beendeleted.
\item \code{NOT_FOUND} if no course exists with the requested ID.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.announcements
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.announcements")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
