% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_announcements_patch}
\alias{gc_courses_announcements_patch}
\title{Announcements Patch}
\usage{
gc_courses_announcements_patch(Announcement, courseId, id, updateMask = NULL)
}
\arguments{
\item{Announcement}{The \link{gc_announcement} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{id}{Identifier of the announcement}

\item{updateMask}{Mask that identifies which fields on the announcement to update}
}
\description{
Updates one or more fields of an announcement.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting developer project did not createthe corresponding announcement or for access errors.
\item \code{INVALID_ARGUMENT} if the request is malformed.
\item \code{FAILED_PRECONDITION} if the requested announcement has already beendeleted.
\item \code{NOT_FOUND} if the requested course or announcement does not exist
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.announcements
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.announcements")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Announcement functions: 
\code{\link{gc_announcement}()},
\code{\link{gc_courses_announcements_create}()}
}
\concept{Announcement functions}
