% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_course_work_delete}
\alias{gc_courses_course_work_delete}
\title{Delete Course Work}
\usage{
gc_courses_course_work_delete(courseId, id)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{id}{Identifier of the course work to delete}
}
\description{
Deletes a course work.
}
\details{
This request must be made by the Developer Console project of the\href{https://support.google.com/cloud/answer/6158849}{OAuth client ID} used tocreate the corresponding course work item. This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting developer project did not createthe corresponding course work, if the requesting user is not permitted to delete the requested course or for access errors.
\itemize{
\item \code{FAILED_PRECONDITION} if the requested course work has already beendeleted.
\item \code{NOT_FOUND} if no course exists with the requested ID.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.students
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.coursework.students")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
