% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_course_work_modify_assignees}
\alias{gc_courses_course_work_modify_assignees}
\title{Modify Assignee}
\usage{
gc_courses_course_work_modify_assignees(
  ModifyCourseWorkAssigneesRequest,
  courseId,
  id
)
}
\arguments{
\item{ModifyCourseWorkAssigneesRequest}{The \link{gc_modify_course_work_assignees_request} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{id}{Identifier of the coursework}
}
\description{
Modifies assignee mode and options of a coursework. Only a teacher of the course that contains the coursework maycall this method.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course or course work or for access errors.
\item \code{INVALID_ARGUMENT} if the request is malformed.
\item \code{NOT_FOUND} if the requested course or course work does not exist.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.students
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.coursework.students")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other ModifyCourseWorkAssigneesRequest functions: 
\code{\link{gc_modify_course_work_assignees_request}()}
}
\concept{ModifyCourseWorkAssigneesRequest functions}
