% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_course_work_patch}
\alias{gc_courses_course_work_patch}
\title{Course Work Patch}
\usage{
gc_courses_course_work_patch(CourseWork, courseId, id, updateMask = NULL)
}
\arguments{
\item{CourseWork}{The \link{gc_course_work} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{id}{Identifier of the course work}

\item{updateMask}{Mask that identifies which fields on the course work to update}
}
\description{
Updates one or more fields of a course work. See google.classroom.v1.CourseWork for detailsof which fields may be updated and who may change them.
This request must be made by the Developer Console project of the \href{https://support.google.com/cloud/answer/6158849}{OAuth client ID} used to
create the corresponding course work item.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course or course work or for access errors.
\item \code{INVALID_ARGUMENT} if the request is malformed.
\item \code{NOT_FOUND} if the requested course or course work does not exist.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.students
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.coursework.students")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other CourseWork functions: 
\code{\link{gc_course_work}()},
\code{\link{gc_courses_course_work_create}()}
}
\concept{CourseWork functions}
