% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_course_work_student_submissions_modify_attachments}
\alias{gc_courses_course_work_student_submissions_modify_attachments}
\title{Modify Attachments}
\usage{
gc_courses_course_work_student_submissions_modify_attachments(
  ModifyAttachmentsRequest,
  courseId,
  courseWorkId,
  id
)
}
\arguments{
\item{ModifyAttachmentsRequest}{The \link{gc_modify_attachments_request} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{courseWorkId}{Identifier of the course work}

\item{id}{Identifier of the student submission}
}
\description{
Modifies attachments of student submission. Attachments may only be added to student submissions belonging to
coursework objects with a \code{workType} of \code{ASSIGNMENT}. This request must be made by the Developer Console
project of the \href{https://support.google.com/cloud/answer/6158849}{OAuth client ID} used tocreate the corresponding
course work item.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course or course work, if the user is not permitted to modifyattachments on the requested student submission, or foraccess errors.
\item \code{INVALID_ARGUMENT} if the request is malformed.
\item \code{NOT_FOUND} if the requested course, course work, or student submissiondoes not exist.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.me
\item https://www.googleapis.com/auth/classroom.coursework.students
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.coursework.me", "https://www.googleapis.com/auth/classroom.coursework.students")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other ModifyAttachmentsRequest functions: 
\code{\link{gc_modify_attachments_request}()}
}
\concept{ModifyAttachmentsRequest functions}
