% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_course_work_student_submissions_patch}
\alias{gc_courses_course_work_student_submissions_patch}
\title{Submissions Patch}
\usage{
gc_courses_course_work_student_submissions_patch(
  StudentSubmission,
  courseId,
  courseWorkId,
  id,
  updateMask = NULL
)
}
\arguments{
\item{StudentSubmission}{The \link{gc_student_submission} object to pass to this method}

\item{courseId}{Identifier of the course}

\item{courseWorkId}{Identifier of the course work}

\item{id}{Identifier of the student submission}

\item{updateMask}{Mask that identifies which fields on the student submission to update}
}
\description{
Updates one or more fields of a student submission. See google.classroom.v1.StudentSubmission
for details of which fields may be updated and who may change them. This request must be made
by the Developer Console project of the \href{https://support.google.com/cloud/answer/6158849}{OAuth client ID}
used tocreate the corresponding course work item.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting developer project did not createthe corresponding course work, if the user is not permitted to make therequested modification to the student submission, or foraccess errors.
\item \code{INVALID_ARGUMENT} if the request is malformed.
\item \code{NOT_FOUND} if the requested course, course work, or student submissiondoes not exist.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.coursework.me
\item https://www.googleapis.com/auth/classroom.coursework.students
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.coursework.me", "https://www.googleapis.com/auth/classroom.coursework.students")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other StudentSubmission functions: 
\code{\link{gc_student_submission}()}
}
\concept{StudentSubmission functions}
