% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_create}
\alias{gc_courses_create}
\title{Create Course}
\usage{
gc_courses_create(Course)
}
\arguments{
\item{Course}{The \link{gc_course} object to pass to this method}
}
\description{
Creates a course
}
\details{
The user specified in \code{ownerId} is the owner of the created courseand added as a teacher.

This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to createcourses or for access errors.
\item \code{NOT_FOUND} if the primary teacher is not a valid user.
\item \code{FAILED_PRECONDITION} if the course owner's account is disabled or forthe following request errors:
\itemize{
\item UserGroupsMembershipLimitReached
}
\item \code{ALREADY_EXISTS} if an alias was specified in the \code{id} andalready exists.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.courses
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.courses")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Course functions: 
\code{\link{gc_courses_patch}()},
\code{\link{gc_courses_update}()},
\code{\link{gc_course}()}
}
\concept{Course functions}
