% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_patch}
\alias{gc_courses_patch}
\title{Courses Patch}
\usage{
gc_courses_patch(Course, id, updateMask = NULL)
}
\arguments{
\item{Course}{The \link{gc_course} object to pass to this method}

\item{id}{Identifier of the course to update}

\item{updateMask}{Mask that identifies which fields on the course to update}
}
\description{
Updates one or more fields in a course.This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to modify therequested course or for access errors.
\item \code{NOT_FOUND} if no course exists with the requested ID.
\item \code{INVALID_ARGUMENT} if invalid fields are specified in the update mask orif no update mask is supplied.
\item \code{FAILED_PRECONDITION} for the following request errors:
\itemize{
\item CourseNotModifiable
}
}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.courses
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.courses")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Course functions: 
\code{\link{gc_courses_create}()},
\code{\link{gc_courses_update}()},
\code{\link{gc_course}()}
}
\concept{Course functions}
