% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_students_create}
\alias{gc_courses_students_create}
\title{Create Students}
\usage{
gc_courses_students_create(Student, courseId, enrollmentCode = NULL)
}
\arguments{
\item{Student}{The \link{gc_student} object to pass to this method}

\item{courseId}{Identifier of the course to create the student in}

\item{enrollmentCode}{Enrollment code of the course to create the student in}
}
\description{
Adds a user as a student of a course.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to createstudents in this course or for access errors.
\item \code{NOT_FOUND} if the requested course ID does not exist.
\item \code{FAILED_PRECONDITION} if the requested user's account is disabled,for the following request errors:
\itemize{
\item CourseMemberLimitReached
\item CourseNotModifiable
\item UserGroupsMembershipLimitReached
}
\item \code{ALREADY_EXISTS} if the user is already a student or teacher in thecourse.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.profile.emails
\item https://www.googleapis.com/auth/classroom.profile.photos
\item https://www.googleapis.com/auth/classroom.rosters
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.profile.emails", "https://www.googleapis.com/auth/classroom.profile.photos", "https://www.googleapis.com/auth/classroom.rosters")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Student functions: 
\code{\link{gc_student}()}
}
\concept{Student functions}
