% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_students_delete}
\alias{gc_courses_students_delete}
\title{Deletes a student of a course}
\usage{
gc_courses_students_delete(courseId, userId)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{userId}{Identifier of the student to delete}
}
\description{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to deletestudents of this course or for access errors.
\item \code{NOT_FOUND} if no student of this course has the requested ID or if thecourse does not exist.
}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.rosters
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.rosters")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
