% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_students_list}
\alias{gc_courses_students_list}
\title{Students List}
\usage{
gc_courses_students_list(courseId, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{pageToken}{nextPageToken}

\item{pageSize}{Maximum number of items to return}
}
\description{
Returns a list of students of this course that the requester is permitted to view.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{NOT_FOUND} if the course does not exist.
\item \code{PERMISSION_DENIED} for access errors.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.profile.emails
\item https://www.googleapis.com/auth/classroom.profile.photos
\item https://www.googleapis.com/auth/classroom.rosters
\item https://www.googleapis.com/auth/classroom.rosters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.profile.emails", "https://www.googleapis.com/auth/classroom.profile.photos", "https://www.googleapis.com/auth/classroom.rosters", "https://www.googleapis.com/auth/classroom.rosters.readonly")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
