% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_teachers_delete}
\alias{gc_courses_teachers_delete}
\title{Delete Teachers}
\usage{
gc_courses_teachers_delete(courseId, userId)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{userId}{Identifier of the teacher to delete}
}
\description{
Deletes a teacher of a course.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to deleteteachers of this course or for access errors.
\item \code{NOT_FOUND} if no teacher of this course has the requested ID or if thecourse does not exist.
\item \code{FAILED_PRECONDITION} if the requested ID belongs to the primary teacherof this course.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.rosters
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.rosters")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
