% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_topics_create}
\alias{gc_courses_topics_create}
\title{Create Topic}
\usage{
gc_courses_topics_create(Topic, courseId)
}
\arguments{
\item{Topic}{The \link{gc_topic} object to pass to this method}

\item{courseId}{Identifier of the course}
}
\description{
Creates a topic
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to access therequested course, create a topic in the requested course,or for access errors.
\item \code{INVALID_ARGUMENT} if the request is malformed.
\item \code{NOT_FOUND} if the requested course does not exist.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.topics
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.topics")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Topic functions: 
\code{\link{gc_courses_topics_patch}()},
\code{\link{gc_topic}()}
}
\concept{Topic functions}
