% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_courses_topics_delete}
\alias{gc_courses_topics_delete}
\title{Delete Topic}
\usage{
gc_courses_topics_delete(courseId, id)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{id}{Identifier of the topic to delete}
}
\description{
Deletes a topic
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not allowed to delete therequested topic or for access errors.
\item \code{FAILED_PRECONDITION} if the requested topic has already beendeleted.
\item \code{NOT_FOUND} if no course or topic exists with the requested ID.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.topics
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.topics")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
