% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_drive_file}
\alias{gc_drive_file}
\title{DriveFile Object}
\usage{
gc_drive_file(
  thumbnailUrl = NULL,
  id = NULL,
  title = NULL,
  alternateLink = NULL
)
}
\arguments{
\item{thumbnailUrl}{URL of a thumbnail image of the Drive item}

\item{id}{Drive API resource ID}

\item{title}{Title of the Drive item}

\item{alternateLink}{URL that can be used to access the Drive item}
}
\value{
DriveFile object
}
\description{
DriveFile Object
}
\details{
Representation of a Google Drive file.
}
\concept{DriveFile functions}
