% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_drive_folder}
\alias{gc_drive_folder}
\title{DriveFolder Object}
\usage{
gc_drive_folder(title = NULL, alternateLink = NULL, id = NULL)
}
\arguments{
\item{title}{Title of the Drive folder}

\item{alternateLink}{URL that can be used to access the Drive folder}

\item{id}{Drive API resource ID}
}
\value{
DriveFolder object
}
\description{
DriveFolder Object
}
\details{
Representation of a Google Drive folder.
}
\concept{DriveFolder functions}
