% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_empty}
\alias{gc_empty}
\title{Empty Object}
\usage{
gc_empty()
}
\value{
Empty object
}
\description{
Empty Object
}
\details{
A generic empty message that you can re-use to avoid defining duplicatedempty messages in your APIs. A typical example is to use it as the requestor the response type of an API method. For instance:    service Foo {      rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);    }The JSON representation for \code{Empty} is empty JSON object \code{{}}.
}
\concept{Empty functions}
