% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_feed}
\alias{gc_feed}
\title{Feed Object}
\usage{
gc_feed(
  courseRosterChangesInfo = NULL,
  courseWorkChangesInfo = NULL,
  feedType = NULL
)
}
\arguments{
\item{courseRosterChangesInfo}{Information about a \code{Feed} with a \code{feed_type} of \code{COURSE_ROSTER_CHANGES}}

\item{courseWorkChangesInfo}{Information about a \code{Feed} with a \code{feed_type} of \code{COURSE_WORK_CHANGES}}

\item{feedType}{The type of feed}
}
\value{
Feed object
}
\description{
Feed Object
}
\details{
A class of notifications that an application can register to receive.For example: 'all roster changes for a domain'.
}
\concept{Feed functions}
