% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_form}
\alias{gc_form}
\title{Form Object}
\usage{
gc_form(title = NULL, thumbnailUrl = NULL, responseUrl = NULL, formUrl = NULL)
}
\arguments{
\item{title}{Title of the Form}

\item{thumbnailUrl}{URL of a thumbnail image of the Form}

\item{responseUrl}{URL of the form responses document}

\item{formUrl}{URL of the form}
}
\value{
Form object
}
\description{
Form Object
}
\details{
Google Forms item.
}
\concept{Form functions}
