% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_global_permission}
\alias{gc_global_permission}
\title{GlobalPermission Object}
\usage{
gc_global_permission(permission = NULL)
}
\arguments{
\item{permission}{Permission value}
}
\value{
GlobalPermission object
}
\description{
GlobalPermission Object
}
\details{
Global user permission description.
}
\concept{GlobalPermission functions}
