% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_grade_history}
\alias{gc_grade_history}
\title{GradeHistory Object}
\usage{
gc_grade_history(
  actorUserId = NULL,
  gradeTimestamp = NULL,
  gradeChangeType = NULL,
  maxPoints = NULL,
  pointsEarned = NULL
)
}
\arguments{
\item{actorUserId}{The teacher who made the grade change}

\item{gradeTimestamp}{When the grade of the submission was changed}

\item{gradeChangeType}{The type of grade change at this time in the submission grade history}

\item{maxPoints}{The denominator of the grade at this time in the submission grade}

\item{pointsEarned}{The numerator of the grade at this time in the submission grade history}
}
\value{
GradeHistory object
}
\description{
GradeHistory Object
}
\details{
The history of each grade on this submission.
}
\concept{GradeHistory functions}
