% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_individual_students_options}
\alias{gc_individual_students_options}
\title{IndividualStudentsOptions Object}
\usage{
gc_individual_students_options(studentIds = NULL)
}
\arguments{
\item{studentIds}{Identifiers for the students that have access to the}
}
\value{
IndividualStudentsOptions object
}
\description{
IndividualStudentsOptions Object
}
\details{
Assignee details about a coursework/announcement.This field is set if and only if \code{assigneeMode} is \code{INDIVIDUAL_STUDENTS}.
}
\concept{IndividualStudentsOptions functions}
