% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_invitation}
\alias{gc_invitation}
\title{Invitation Object}
\usage{
gc_invitation(id = NULL, role = NULL, userId = NULL, courseId = NULL)
}
\arguments{
\item{id}{Identifier assigned by Classroom}

\item{role}{Role to invite the user to have}

\item{userId}{Identifier of the invited user}

\item{courseId}{Identifier of the course to invite the user to}
}
\value{
Invitation object
}
\description{
Invitation Object
}
\details{
An invitation to join a course.
}
\seealso{
Other Invitation functions: 
\code{\link{gc_invitations_create}()}
}
\concept{Invitation functions}
