% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_invitations_accept}
\alias{gc_invitations_accept}
\title{Accepts an invitation}
\usage{
gc_invitations_accept(id)
}
\arguments{
\item{id}{Identifier of the invitation to accept}
}
\description{
Accepts an invitation, removing it and adding the invited user to theteachers or students (as appropriate) of
the specified course. Only theinvited user may accept an invitation. This method returns the following error
codes:
}
\details{
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to accept therequested invitation or for access errors.
\item \code{FAILED_PRECONDITION} for the following request errors:
\itemize{
\item CourseMemberLimitReached
\item CourseNotModifiable
\item CourseTeacherLimitReached
\item UserGroupsMembershipLimitReached
}
\item \code{NOT_FOUND} if no invitation exists with the requested ID.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.rosters
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.rosters")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
