% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_invitations_create}
\alias{gc_invitations_create}
\title{Creates an invitation}
\usage{
gc_invitations_create(Invitation)
}
\arguments{
\item{Invitation}{The \link{gc_invitation} object to pass to this method}
}
\description{
Only one invitation for a user and course may existat a time. Delete and re-create an invitation to
make changes.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to createinvitations for this course or for access errors.
\item \code{NOT_FOUND} if the course or the user does not exist.
\item \code{FAILED_PRECONDITION} if the requested user's account is disabled or ifthe user already has this role or a role with greater permissions.
\item \code{ALREADY_EXISTS} if an invitation for the specified user and coursealready exists.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.rosters
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.rosters")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Invitation functions: 
\code{\link{gc_invitation}()}
}
\concept{Invitation functions}
