% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_invitations_delete}
\alias{gc_invitations_delete}
\title{Deletes an invitation}
\usage{
gc_invitations_delete(id)
}
\arguments{
\item{id}{Identifier of the invitation to delete}
}
\description{
This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to delete therequested invitation or for access errors.* \code{NOT_FOUND} if no invitation exists with the requested ID.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.rosters
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.rosters")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
