% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_invitations_get}
\alias{gc_invitations_get}
\title{Returns an invitation}
\usage{
gc_invitations_get(id)
}
\arguments{
\item{id}{Identifier of the invitation to return}
}
\description{
This method returns the following error codes:
}
\details{
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to view therequested invitation or for access errors.
\item \code{NOT_FOUND} if no invitation exists with the requested ID.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.rosters
\item https://www.googleapis.com/auth/classroom.rosters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.rosters", "https://www.googleapis.com/auth/classroom.rosters.readonly")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
