% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_link}
\alias{gc_link}
\title{Link Object}
\usage{
gc_link(title = NULL, thumbnailUrl = NULL, url = NULL)
}
\arguments{
\item{title}{Title of the target of the URL}

\item{thumbnailUrl}{URL of a thumbnail image of the target URL}

\item{url}{URL to link to}
}
\value{
Link object
}
\description{
Link Object
}
\details{
URL item.
}
\concept{Link functions}
