% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_list_announcements_response}
\alias{gc_list_announcements_response}
\title{ListAnnouncementsResponse Object}
\usage{
gc_list_announcements_response(announcements = NULL, nextPageToken = NULL)
}
\arguments{
\item{announcements}{Announcement items that match the request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListAnnouncementsResponse object
}
\description{
ListAnnouncementsResponse Object
}
\details{
Response when listing course work.
}
\concept{ListAnnouncementsResponse functions}
