% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_list_course_aliases_response}
\alias{gc_list_course_aliases_response}
\title{ListCourseAliasesResponse Object}
\usage{
gc_list_course_aliases_response(nextPageToken = NULL, aliases = NULL)
}
\arguments{
\item{nextPageToken}{Token identifying the next page of results to return}

\item{aliases}{The course aliases}
}
\value{
ListCourseAliasesResponse object
}
\description{
ListCourseAliasesResponse Object
}
\details{
Response when listing course aliases.
}
\concept{ListCourseAliasesResponse functions}
