% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_list_course_work_response}
\alias{gc_list_course_work_response}
\title{ListCourseWorkResponse Object}
\usage{
gc_list_course_work_response(nextPageToken = NULL, courseWork = NULL)
}
\arguments{
\item{nextPageToken}{Token identifying the next page of results to return}

\item{courseWork}{Course work items that match the request}
}
\value{
ListCourseWorkResponse object
}
\description{
ListCourseWorkResponse Object
}
\details{
Response when listing course work.
}
\concept{ListCourseWorkResponse functions}
