% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_list_courses_response}
\alias{gc_list_courses_response}
\title{ListCoursesResponse Object}
\usage{
gc_list_courses_response(nextPageToken = NULL, courses = NULL)
}
\arguments{
\item{nextPageToken}{Token identifying the next page of results to return}

\item{courses}{Courses that match the list request}
}
\value{
ListCoursesResponse object
}
\description{
ListCoursesResponse Object
}
\details{
Response when listing courses.
}
\concept{ListCoursesResponse functions}
