% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_list_guardian_invitations_response}
\alias{gc_list_guardian_invitations_response}
\title{ListGuardianInvitationsResponse Object}
\usage{
gc_list_guardian_invitations_response(
  guardianInvitations = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{guardianInvitations}{Guardian invitations that matched the list request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListGuardianInvitationsResponse object
}
\description{
ListGuardianInvitationsResponse Object
}
\details{
Response when listing guardian invitations.
}
\concept{ListGuardianInvitationsResponse functions}
