% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_list_guardians_response}
\alias{gc_list_guardians_response}
\title{ListGuardiansResponse Object}
\usage{
gc_list_guardians_response(nextPageToken = NULL, guardians = NULL)
}
\arguments{
\item{nextPageToken}{Token identifying the next page of results to return}

\item{guardians}{Guardians on this page of results that met the criteria specified in}
}
\value{
ListGuardiansResponse object
}
\description{
ListGuardiansResponse Object
}
\details{
Response when listing guardians.
}
\concept{ListGuardiansResponse functions}
