% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_list_invitations_response}
\alias{gc_list_invitations_response}
\title{ListInvitationsResponse Object}
\usage{
gc_list_invitations_response(nextPageToken = NULL, invitations = NULL)
}
\arguments{
\item{nextPageToken}{Token identifying the next page of results to return}

\item{invitations}{Invitations that match the list request}
}
\value{
ListInvitationsResponse object
}
\description{
ListInvitationsResponse Object
}
\details{
Response when listing invitations.
}
\concept{ListInvitationsResponse functions}
