% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_list_student_submissions_response}
\alias{gc_list_student_submissions_response}
\title{ListStudentSubmissionsResponse Object}
\usage{
gc_list_student_submissions_response(
  studentSubmissions = NULL,
  nextPageToken = NULL
)
}
\arguments{
\item{studentSubmissions}{Student work that matches the request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListStudentSubmissionsResponse object
}
\description{
ListStudentSubmissionsResponse Object
}
\details{
Response when listing student submissions.
}
\concept{ListStudentSubmissionsResponse functions}
