% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_list_students_response}
\alias{gc_list_students_response}
\title{ListStudentsResponse Object}
\usage{
gc_list_students_response(students = NULL, nextPageToken = NULL)
}
\arguments{
\item{students}{Students who match the list request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListStudentsResponse object
}
\description{
ListStudentsResponse Object
}
\details{
Response when listing students.
}
\concept{ListStudentsResponse functions}
