% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_list_teachers_response}
\alias{gc_list_teachers_response}
\title{Google Classroom API Objects
Manages classes, rosters, and invitations in Google Classroom.
ListTeachersResponse Object}
\usage{
gc_list_teachers_response(teachers = NULL, nextPageToken = NULL)
}
\arguments{
\item{teachers}{Teachers who match the list request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListTeachersResponse object
}
\description{
Google Classroom API Objects
Manages classes, rosters, and invitations in Google Classroom.
ListTeachersResponse Object
}
\details{
Response when listing teachers.
}
\concept{ListTeachersResponse functions}
