% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_lst_topic_response}
\alias{gc_lst_topic_response}
\title{ListTopicResponse Object}
\usage{
gc_lst_topic_response(nextPageToken = NULL, topic = NULL)
}
\arguments{
\item{nextPageToken}{Token identifying the next page of results to return}

\item{topic}{Topic items that match the request}
}
\value{
ListTopicResponse object
}
\description{
ListTopicResponse Object
}
\details{
Response when listing topics.
}
\concept{ListTopicResponse functions}
