% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_material}
\alias{gc_material}
\title{Material Object}
\usage{
gc_material(youtubeVideo = NULL, driveFile = NULL, form = NULL, link = NULL)
}
\arguments{
\item{youtubeVideo}{YouTube video material}

\item{driveFile}{Google Drive file material}

\item{form}{Google Forms material}

\item{link}{Link material}
}
\value{
Material object
}
\description{
Material Object
}
\details{
Material attached to course work.When creating attachments, setting the \code{form} field is not supported.
}
\concept{Material functions}
