% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_modify_announcement_assignees_request}
\alias{gc_modify_announcement_assignees_request}
\title{ModifyAnnouncementAssigneesRequest Object}
\usage{
gc_modify_announcement_assignees_request(
  modifyIndividualStudentsOptions = NULL,
  assigneeMode = NULL
)
}
\arguments{
\item{modifyIndividualStudentsOptions}{Set which students can view or cannot view the announcement}

\item{assigneeMode}{Mode of the announcement describing whether it is accessible by all}
}
\value{
ModifyAnnouncementAssigneesRequest object
}
\description{
ModifyAnnouncementAssigneesRequest Object
}
\details{
Request to modify assignee mode and options of an announcement.
}
\seealso{
Other ModifyAnnouncementAssigneesRequest functions: 
\code{\link{gc_courses_announcements_modify_assignees}()}
}
\concept{ModifyAnnouncementAssigneesRequest functions}
