% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_modify_attachments_request}
\alias{gc_modify_attachments_request}
\title{ModifyAttachmentsRequest Object}
\usage{
gc_modify_attachments_request(addAttachments = NULL)
}
\arguments{
\item{addAttachments}{Attachments to add}
}
\value{
ModifyAttachmentsRequest object
}
\description{
ModifyAttachmentsRequest Object
}
\details{
Request to modify the attachments of a student submission.
}
\seealso{
Other ModifyAttachmentsRequest functions: 
\code{\link{gc_courses_course_work_student_submissions_modify_attachments}()}
}
\concept{ModifyAttachmentsRequest functions}
