% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_modify_course_work_assignees_request}
\alias{gc_modify_course_work_assignees_request}
\title{ModifyCourseWorkAssigneesRequest Object}
\usage{
gc_modify_course_work_assignees_request(
  modifyIndividualStudentsOptions = NULL,
  assigneeMode = NULL
)
}
\arguments{
\item{modifyIndividualStudentsOptions}{Set which students are assigned or not assigned to the coursework}

\item{assigneeMode}{Mode of the coursework describing whether it will be assigned to all}
}
\value{
ModifyCourseWorkAssigneesRequest object
}
\description{
ModifyCourseWorkAssigneesRequest Object
}
\details{
Request to modify assignee mode and options of a coursework.
}
\seealso{
Other ModifyCourseWorkAssigneesRequest functions: 
\code{\link{gc_courses_course_work_modify_assignees}()}
}
\concept{ModifyCourseWorkAssigneesRequest functions}
