% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_modify_individual_students_options}
\alias{gc_modify_individual_students_options}
\title{ModifyIndividualStudentsOptions Object}
\usage{
gc_modify_individual_students_options(
  addStudentIds = NULL,
  removeStudentIds = NULL
)
}
\arguments{
\item{addStudentIds}{IDs of students to be added as having access to this}

\item{removeStudentIds}{IDs of students to be removed from having access to this}
}
\value{
ModifyIndividualStudentsOptions object
}
\description{
ModifyIndividualStudentsOptions Object
}
\details{
Contains fields to add or remove students from a course work or announcementwhere the \code{assigneeMode} is set to \code{INDIVIDUAL_STUDENTS}.
}
\concept{ModifyIndividualStudentsOptions functions}
