% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_multiple_choice_question}
\alias{gc_multiple_choice_question}
\title{MultipleChoiceQuestion Object}
\usage{
gc_multiple_choice_question(choices = NULL)
}
\arguments{
\item{choices}{Possible choices}
}
\value{
MultipleChoiceQuestion object
}
\description{
MultipleChoiceQuestion Object
}
\details{
Additional details for multiple-choice questions.
}
\concept{MultipleChoiceQuestion functions}
