% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_multiple_choice_submission}
\alias{gc_multiple_choice_submission}
\title{MultipleChoiceSubmission Object}
\usage{
gc_multiple_choice_submission(answer = NULL)
}
\arguments{
\item{answer}{Student's select choice}
}
\value{
MultipleChoiceSubmission object
}
\description{
MultipleChoiceSubmission Object
}
\details{
Student work for a multiple-choice question.
}
\concept{MultipleChoiceSubmission functions}
