% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_name}
\alias{gc_name}
\title{Name Object}
\usage{
gc_name(givenName = NULL, familyName = NULL, fullName = NULL)
}
\arguments{
\item{givenName}{The user's first name}

\item{familyName}{The user's last name}

\item{fullName}{The user's full name formed by concatenating the first and last name}
}
\value{
Name object
}
\description{
Name Object
}
\details{
Details of the user's name.
}
\concept{Name functions}
