% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_reclaim_student_submission_request}
\alias{gc_reclaim_student_submission_request}
\title{ReclaimStudentSubmissionRequest Object}
\usage{
gc_reclaim_student_submission_request()
}
\value{
ReclaimStudentSubmissionRequest object
}
\description{
ReclaimStudentSubmissionRequest Object
}
\details{
Request to reclaim a student submission.
}
\seealso{
Other ReclaimStudentSubmissionRequest functions: 
\code{\link{gc_courses_course_work_student_submissions_reclaim}()}
}
\concept{ReclaimStudentSubmissionRequest functions}
