% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_return_student_submission_request}
\alias{gc_return_student_submission_request}
\title{ReturnStudentSubmissionRequest Object}
\usage{
gc_return_student_submission_request()
}
\value{
ReturnStudentSubmissionRequest object
}
\description{
ReturnStudentSubmissionRequest Object
}
\details{
Request to return a student submission.
}
\seealso{
Other ReturnStudentSubmissionRequest functions: 
\code{\link{gc_courses_course_work_student_submissions_return}()}
}
\concept{ReturnStudentSubmissionRequest functions}
