% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_shared_drive_file}
\alias{gc_shared_drive_file}
\title{SharedDriveFile Object}
\usage{
gc_shared_drive_file(driveFile = NULL, shareMode = NULL)
}
\arguments{
\item{driveFile}{Drive file details}

\item{shareMode}{Mechanism by which students access the Drive item}
}
\value{
SharedDriveFile object
}
\description{
SharedDriveFile Object
}
\details{
Drive file that is used as material for course work.
}
\concept{SharedDriveFile functions}
