% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_short_answer_submission}
\alias{gc_short_answer_submission}
\title{ShortAnswerSubmission Object}
\usage{
gc_short_answer_submission(answer = NULL)
}
\arguments{
\item{answer}{Student response to a short-answer question}
}
\value{
ShortAnswerSubmission object
}
\description{
ShortAnswerSubmission Object
}
\details{
Student work for a short answer question.
}
\concept{ShortAnswerSubmission functions}
