% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_student}
\alias{gc_student}
\title{Student Object}
\usage{
gc_student(
  userId = NULL,
  courseId = NULL,
  profile = NULL,
  studentWorkFolder = NULL
)
}
\arguments{
\item{userId}{Identifier of the user}

\item{courseId}{Identifier of the course}

\item{profile}{Global user information for the student}

\item{studentWorkFolder}{Information about a Drive Folder for this student's work in this course}
}
\value{
Student object
}
\description{
Student Object
}
\details{
Student in a course.
}
\seealso{
Other Student functions: 
\code{\link{gc_courses_students_create}()}
}
\concept{Student functions}
