% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_submission_history}
\alias{gc_submission_history}
\title{SubmissionHistory Object}
\usage{
gc_submission_history(stateHistory = NULL, gradeHistory = NULL)
}
\arguments{
\item{stateHistory}{The state history information of the submission, if present}

\item{gradeHistory}{The grade history information of the submission, if present}
}
\value{
SubmissionHistory object
}
\description{
SubmissionHistory Object
}
\details{
The history of the submission. This currently includes state and gradehistories.
}
\concept{SubmissionHistory functions}
