% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_teacher}
\alias{gc_teacher}
\title{Teacher Object}
\usage{
gc_teacher(profile = NULL, userId = NULL, courseId = NULL)
}
\arguments{
\item{profile}{Global user information for the teacher}

\item{userId}{Identifier of the user}

\item{courseId}{Identifier of the course}
}
\value{
Teacher object
}
\description{
Teacher Object
}
\details{
Teacher of a course.
}
\seealso{
Other Teacher functions: 
\code{\link{gc_courses_teachers_create}()}
}
\concept{Teacher functions}
