% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_time_of_day}
\alias{gc_time_of_day}
\title{TimeOfDay Object}
\usage{
gc_time_of_day(seconds = NULL, minutes = NULL, hours = NULL, nanos = NULL)
}
\arguments{
\item{seconds}{Seconds of minutes of the time}

\item{minutes}{Minutes of hour of day}

\item{hours}{Hours of day in 24 hour format}

\item{nanos}{Fractions of seconds in nanoseconds}
}
\value{
TimeOfDay object
}
\description{
TimeOfDay Object
}
\details{
Represents a time of day. The date and time zone are either not significantor are specified elsewhere. An API may choose to allow leap seconds. Relatedtypes are google.type.Date and \code{google.protobuf.Timestamp}.
}
\concept{TimeOfDay functions}
