% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_topic}
\alias{gc_topic}
\title{Topic Object}
\usage{
gc_topic(courseId = NULL, updateTime = NULL, name = NULL, topicId = NULL)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{updateTime}{The time the topic was last updated by the system}

\item{name}{The name of the topic, generated by the user}

\item{topicId}{Unique identifier for the topic}
}
\value{
Topic object
}
\description{
Topic Object
}
\details{
Topic created by a teacher for the course
}
\seealso{
Other Topic functions: 
\code{\link{gc_courses_topics_create}()},
\code{\link{gc_courses_topics_patch}()}
}
\concept{Topic functions}
