% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_turn_in_student_submission_request}
\alias{gc_turn_in_student_submission_request}
\title{TurnInStudentSubmissionRequest Object}
\usage{
gc_turn_in_student_submission_request()
}
\value{
TurnInStudentSubmissionRequest object
}
\description{
TurnInStudentSubmissionRequest Object
}
\details{
Request to turn in a student submission.
}
\seealso{
Other TurnInStudentSubmissionRequest functions: 
\code{\link{gc_courses_course_work_student_submissions_turn_in}()}
}
\concept{TurnInStudentSubmissionRequest functions}
