% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_user_profiles_guardian_invitations_create}
\alias{gc_user_profiles_guardian_invitations_create}
\title{Create Guardian Invitations}
\usage{
gc_user_profiles_guardian_invitations_create(GuardianInvitation, studentId)
}
\arguments{
\item{GuardianInvitation}{The \link{gc_guardian_invitation} object to pass to this method}

\item{studentId}{ID of the student (in standard format)}
}
\description{
Creates a guardian invitation, and sends an email to the guardian askingthem to confirm that they are the
student's guardian. Once the guardian accepts the invitation, their \code{state} will change to\code{COMPLETED} and
they will start receiving guardian notifications. A\code{Guardian} resource will also be created to represent
the active guardian. The request object must have the \code{student_id} and\code{invited_email_address} fields set.
Failing to set these fields, orsetting any other fields in the request, will result in an error.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the current user does not have permission to  manage guardians, if the guardian in question has already rejected  too many requests for that student, if guardians are not enabled for the  domain in question, or for other access errors.
\item \code{RESOURCE_EXHAUSTED} if the student or guardian has exceeded the guardian  link limit.
\item \code{INVALID_ARGUMENT} if the guardian email address is not valid (for  example, if it is too long), or if the format of the student ID provided  cannot be recognized (it is not an email address, nor a \code{user_id} from  this API). This error will also be returned if read-only fields are set,  or if the \code{state} field is set to to a value other than \code{PENDING}.
\item \code{NOT_FOUND} if the student ID provided is a valid student ID, but  Classroom has no record of that student.
\item \code{ALREADY_EXISTS} if there is already a pending guardian invitation for  the student and \code{invited_email_address} provided, or if the provided  \code{invited_email_address} matches the Google account of an existing  \code{Guardian} for this user.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.guardianlinks.students
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.guardianlinks.students")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other GuardianInvitation functions: 
\code{\link{gc_guardian_invitation}()},
\code{\link{gc_user_profiles_guardian_invitations_patch}()}
}
\concept{GuardianInvitation functions}
