% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_user_profiles_guardian_invitations_get}
\alias{gc_user_profiles_guardian_invitations_get}
\title{Get Guardian Invitations}
\usage{
gc_user_profiles_guardian_invitations_get(studentId, invitationId)
}
\arguments{
\item{studentId}{The ID of the student whose guardian invitation is being requested}

\item{invitationId}{The \code{id} field of the \code{GuardianInvitation} being requested}
}
\description{
Returns a specific guardian invitation.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if the requesting user is not permitted to view  guardian invitations for the student identified by the \code{student_id}, if  guardians are not enabled for the domain in question, or for other  access errors.
\item \code{INVALID_ARGUMENT} if a \code{student_id} is specified, but its format cannot  be recognized (it is not an email address, nor a \code{student_id} from the  API, nor the literal string \code{me}).
\item \code{NOT_FOUND} if Classroom cannot find any record of the given student or  \code{invitation_id}. May also be returned if the student exists, but the  requesting user does not have access to see that student.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.guardianlinks.students
\item https://www.googleapis.com/auth/classroom.guardianlinks.students.readonly
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.guardianlinks.students", "https://www.googleapis.com/auth/classroom.guardianlinks.students.readonly")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
