% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_user_profiles_guardian_invitations_list}
\alias{gc_user_profiles_guardian_invitations_list}
\title{Guardian Invitations List}
\usage{
gc_user_profiles_guardian_invitations_list(
  studentId,
  pageToken = NULL,
  invitedEmailAddress = NULL,
  states = NULL,
  pageSize = NULL
)
}
\arguments{
\item{studentId}{The ID of the student whose guardian invitations are to be returned}

\item{pageToken}{nextPageToken}

\item{invitedEmailAddress}{If specified, only results with the specified \code{invited_email_address}}

\item{states}{If specified, only results with the specified \code{state} values are}

\item{pageSize}{Maximum number of items to return}
}
\description{
Returns a list of guardian invitations that the requesting user ispermitted to view, filtered by the
parameters provided.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if a \code{student_id} is specified, and the requesting  user is not permitted to view guardian invitations for that student, if  \code{'-'} is specified as the \code{student_id} and the user is not a domain  administrator, if guardians are not enabled for the domain in question,  or for other access errors.
\item \code{INVALID_ARGUMENT} if a \code{student_id} is specified, but its format cannot  be recognized (it is not an email address, nor a \code{student_id} from the  API, nor the literal string \code{me}). May also be returned if an invalid  \code{page_token} or \code{state} is provided.
\item \code{NOT_FOUND} if a \code{student_id} is specified, and its format can be  recognized, but Classroom has no record of that student.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.guardianlinks.students
\item https://www.googleapis.com/auth/classroom.guardianlinks.students.readonly
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.guardianlinks.students", "https://www.googleapis.com/auth/classroom.guardianlinks.students.readonly")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
