% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{gc_user_profiles_guardians_get}
\alias{gc_user_profiles_guardians_get}
\title{Get Guardians}
\usage{
gc_user_profiles_guardians_get(studentId, guardianId)
}
\arguments{
\item{studentId}{The student whose guardian is being requested}

\item{guardianId}{The \code{id} field from a \code{Guardian}}
}
\description{
Returns a specific guardian.
}
\details{
This method returns the following error codes:
\itemize{
\item \code{PERMISSION_DENIED} if no user that matches the provided \code{student_id}  is visible to the requesting user, if the requesting user is not  permitted to view guardian information for the student identified by the  \code{student_id}, if guardians are not enabled for the domain in question,  or for other access errors.
\item \code{INVALID_ARGUMENT} if a \code{student_id} is specified, but its format cannot  be recognized (it is not an email address, nor a \code{student_id} from the  API, nor the literal string \code{me}).
\item \code{NOT_FOUND} if the requesting user is permitted to view guardians for  the requested \code{student_id}, but no \code{Guardian} record exists for that  student that matches the provided \code{guardian_id}.
}

Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.guardianlinks.me.readonly
\item https://www.googleapis.com/auth/classroom.guardianlinks.students
\item https://www.googleapis.com/auth/classroom.guardianlinks.students.readonly
}

Set \code{options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.guardianlinks.me.readonly", "https://www.googleapis.com/auth/classroom.guardianlinks.students", "https://www.googleapis.com/auth/classroom.guardianlinks.students.readonly")}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
