% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{gc_youtube_video}
\alias{gc_youtube_video}
\title{YouTubeVideo Object}
\usage{
gc_youtube_video(
  id = NULL,
  title = NULL,
  alternateLink = NULL,
  thumbnailUrl = NULL
)
}
\arguments{
\item{id}{YouTube API resource ID}

\item{title}{Title of the YouTube video}

\item{alternateLink}{URL that can be used to view the YouTube video}

\item{thumbnailUrl}{URL of a thumbnail image of the YouTube video}
}
\value{
YouTubeVideo object
}
\description{
YouTubeVideo Object
}
\details{
YouTube video item.
}
\concept{YouTubeVideo functions}
