% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursework.R
\name{get_studentSubmissionForACourseWork}
\alias{get_studentSubmissionForACourseWork}
\title{Get student submission for a course work}
\usage{
get_studentSubmissionForACourseWork(courseWorkId, roster)
}
\arguments{
\item{courseWorkId}{A character. can be chosen from the return of list_courseworksAsDataframe}

\item{roster}{A data frame which has 姓名 學號 googleUserId}
}
\value{

}
\description{
Get student submission for a course work
}
\examples{
none
}
