% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Guardian}
\alias{Guardian}
\title{Guardian Object}
\usage{
Guardian(
  guardianId = NULL,
  invitedEmailAddress = NULL,
  guardianProfile = NULL,
  studentId = NULL
)
}
\arguments{
\item{guardianId}{Identifier for the guardian}

\item{invitedEmailAddress}{The email address to which the initial guardian invitation was sent}

\item{guardianProfile}{User profile for the guardian}

\item{studentId}{Identifier for the student to whom the guardian relationship applies}
}
\value{
Guardian object
}
\description{
Guardian Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Association between a student and a guardian of that student. The guardianmay receive information about the student's course work.
}
\concept{Guardian functions}
