% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{GuardianInvitation}
\alias{GuardianInvitation}
\title{GuardianInvitation Object}
\usage{
GuardianInvitation(
  studentId = NULL,
  state = NULL,
  invitedEmailAddress = NULL,
  creationTime = NULL,
  invitationId = NULL
)
}
\arguments{
\item{studentId}{ID of the student (in standard format)}

\item{state}{The state that this invitation is in}

\item{invitedEmailAddress}{Email address that the invitation was sent to}

\item{creationTime}{The time that this invitation was created}

\item{invitationId}{Unique identifier for this invitation}
}
\value{
GuardianInvitation object
}
\description{
GuardianInvitation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An invitation to become the guardian of a specified user, sent to a specifiedemail address.
}
\seealso{
Other GuardianInvitation functions: 
\code{\link{userProfiles.guardianInvitations.create}()},
\code{\link{userProfiles.guardianInvitations.patch}()}
}
\concept{GuardianInvitation functions}
