% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{IndividualStudentsOptions}
\alias{IndividualStudentsOptions}
\title{IndividualStudentsOptions Object}
\usage{
IndividualStudentsOptions(studentIds = NULL)
}
\arguments{
\item{studentIds}{Identifiers for the students that have access to the}
}
\value{
IndividualStudentsOptions object
}
\description{
IndividualStudentsOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Assignee details about a coursework/announcement.This field is set if and only if \code{assigneeMode} is \code{INDIVIDUAL_STUDENTS}.
}
\concept{IndividualStudentsOptions functions}
