% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{invitations.create}
\alias{invitations.create}
\title{Creates an invitation. Only one invitation for a user and course may existat a time. Delete and re-create an invitation to make changes.This method returns the following error codes:* \code{PERMISSION_DENIED} if the requesting user is not permitted to createinvitations for this course or for access errors.* \code{NOT_FOUND} if the course or the user does not exist.* \code{FAILED_PRECONDITION} if the requested user's account is disabled or ifthe user already has this role or a role with greater permissions.* \code{ALREADY_EXISTS} if an invitation for the specified user and coursealready exists.}
\usage{
invitations.create(Invitation)
}
\arguments{
\item{Invitation}{The \link{Invitation} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.rosters
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.rosters)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other Invitation functions: 
\code{\link{Invitation}()}
}
\concept{Invitation functions}
