% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{invitations.list}
\alias{invitations.list}
\title{Returns a list of invitations that the requesting user is permitted toview, restricted to those that match the list request.\emph{Note:} At least one of \code{user_id} or \code{course_id} must be supplied. Bothfields can be supplied.This method returns the following error codes:* \code{PERMISSION_DENIED} for access errors.}
\usage{
invitations.list(
  courseId = NULL,
  userId = NULL,
  pageToken = NULL,
  pageSize = NULL
)
}
\arguments{
\item{courseId}{Restricts returned invitations to those for a course with the specified}

\item{userId}{Restricts returned invitations to those for a specific user}

\item{pageToken}{nextPageToken}

\item{pageSize}{Maximum number of items to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.rosters
\item https://www.googleapis.com/auth/classroom.rosters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.rosters, https://www.googleapis.com/auth/classroom.rosters.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
