% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursework.R
\name{list_courseworksAsDataframe}
\alias{list_courseworksAsDataframe}
\title{list all course works}
\usage{
list_courseworksAsDataframe(courseId)
}
\arguments{
\item{courseId}{}
}
\value{

}
\description{
list all course works
}
\examples{
none.
}
