% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListAnnouncementsResponse}
\alias{ListAnnouncementsResponse}
\title{ListAnnouncementsResponse Object}
\usage{
ListAnnouncementsResponse(announcements = NULL, nextPageToken = NULL)
}
\arguments{
\item{announcements}{Announcement items that match the request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListAnnouncementsResponse object
}
\description{
ListAnnouncementsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing course work.
}
\concept{ListAnnouncementsResponse functions}
