% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListInvitationsResponse}
\alias{ListInvitationsResponse}
\title{ListInvitationsResponse Object}
\usage{
ListInvitationsResponse(nextPageToken = NULL, invitations = NULL)
}
\arguments{
\item{nextPageToken}{Token identifying the next page of results to return}

\item{invitations}{Invitations that match the list request}
}
\value{
ListInvitationsResponse object
}
\description{
ListInvitationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing invitations.
}
\concept{ListInvitationsResponse functions}
