% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListTeachersResponse}
\alias{ListTeachersResponse}
\title{Google Classroom API Objects
Manages classes, rosters, and invitations in Google Classroom.}
\usage{
ListTeachersResponse(teachers = NULL, nextPageToken = NULL)
}
\arguments{
\item{teachers}{Teachers who match the list request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListTeachersResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
at 2020-05-01 22:10:36
filename: ..//googleclassroomv1.auto/R/classroom_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListTeachersResponse Object

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing teachers.
}
\concept{ListTeachersResponse functions}
