% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListTopicResponse}
\alias{ListTopicResponse}
\title{ListTopicResponse Object}
\usage{
ListTopicResponse(nextPageToken = NULL, topic = NULL)
}
\arguments{
\item{nextPageToken}{Token identifying the next page of results to return}

\item{topic}{Topic items that match the request}
}
\value{
ListTopicResponse object
}
\description{
ListTopicResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing topics.
}
\concept{ListTopicResponse functions}
