% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ModifyCourseWorkAssigneesRequest}
\alias{ModifyCourseWorkAssigneesRequest}
\title{ModifyCourseWorkAssigneesRequest Object}
\usage{
ModifyCourseWorkAssigneesRequest(
  modifyIndividualStudentsOptions = NULL,
  assigneeMode = NULL
)
}
\arguments{
\item{modifyIndividualStudentsOptions}{Set which students are assigned or not assigned to the coursework}

\item{assigneeMode}{Mode of the coursework describing whether it will be assigned to all}
}
\value{
ModifyCourseWorkAssigneesRequest object
}
\description{
ModifyCourseWorkAssigneesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to modify assignee mode and options of a coursework.
}
\seealso{
Other ModifyCourseWorkAssigneesRequest functions: 
\code{\link{courses.courseWork.modifyAssignees}()}
}
\concept{ModifyCourseWorkAssigneesRequest functions}
