% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Registration}
\alias{Registration}
\title{Registration Object}
\usage{
Registration(
  feed = NULL,
  registrationId = NULL,
  cloudPubsubTopic = NULL,
  expiryTime = NULL
)
}
\arguments{
\item{feed}{Specification for the class of notifications that Classroom should deliver}

\item{registrationId}{A server-generated unique identifier for this \code{Registration}}

\item{cloudPubsubTopic}{The Cloud Pub/Sub topic that notifications are to be sent to}

\item{expiryTime}{The time until which the \code{Registration} is effective}
}
\value{
Registration object
}
\description{
Registration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An instruction to Classroom to send notifications from the \code{feed} to theprovided destination.
}
\seealso{
Other Registration functions: 
\code{\link{registrations.create}()}
}
\concept{Registration functions}
