% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{registrations.delete}
\alias{registrations.delete}
\title{Deletes a \code{Registration}, causing Classroom to stop sending notificationsfor that \code{Registration}.}
\usage{
registrations.delete(registrationId)
}
\arguments{
\item{registrationId}{The \code{registration_id} of the \code{Registration} to be deleted}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.push-notifications
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.push-notifications)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
