% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{StateHistory}
\alias{StateHistory}
\title{StateHistory Object}
\usage{
StateHistory(state = NULL, stateTimestamp = NULL, actorUserId = NULL)
}
\arguments{
\item{state}{The workflow pipeline stage}

\item{stateTimestamp}{When the submission entered this state}

\item{actorUserId}{The teacher or student who made the change}
}
\value{
StateHistory object
}
\description{
StateHistory Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The history of each state this submission has been in.
}
\concept{StateHistory functions}
