% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{StudentSubmission}
\alias{StudentSubmission}
\title{StudentSubmission Object}
\usage{
StudentSubmission(
  multipleChoiceSubmission = NULL,
  assignmentSubmission = NULL,
  associatedWithDeveloper = NULL,
  shortAnswerSubmission = NULL,
  updateTime = NULL,
  alternateLink = NULL,
  draftGrade = NULL,
  late = NULL,
  courseWorkType = NULL,
  creationTime = NULL,
  state = NULL,
  userId = NULL,
  courseWorkId = NULL,
  courseId = NULL,
  id = NULL,
  submissionHistory = NULL,
  assignedGrade = NULL
)
}
\arguments{
\item{multipleChoiceSubmission}{Submission content when course_work_type is MULTIPLE_CHOICE_QUESTION}

\item{assignmentSubmission}{Submission content when course_work_type is ASSIGNMENT}

\item{associatedWithDeveloper}{Whether this student submission is associated with the Developer Console}

\item{shortAnswerSubmission}{Submission content when course_work_type is SHORT_ANSWER_QUESTION}

\item{updateTime}{Last update time of this submission}

\item{alternateLink}{Absolute link to the submission in the Classroom web UI}

\item{draftGrade}{Optional pending grade}

\item{late}{Whether this submission is late}

\item{courseWorkType}{Type of course work this submission is for}

\item{creationTime}{Creation time of this submission}

\item{state}{State of this submission}

\item{userId}{Identifier for the student that owns this submission}

\item{courseWorkId}{Identifier for the course work this corresponds to}

\item{courseId}{Identifier of the course}

\item{id}{Classroom-assigned Identifier for the student submission}

\item{submissionHistory}{The history of the submission (includes state and grade histories)}

\item{assignedGrade}{Optional grade}
}
\value{
StudentSubmission object
}
\description{
StudentSubmission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Student submission for course work.StudentSubmission items are generated when a CourseWork item is created.StudentSubmissions that have never been accessed (i.e. with \code{state} = NEW)may not have a creation time or update time.
}
\seealso{
Other StudentSubmission functions: 
\code{\link{courses.courseWork.studentSubmissions.patch}()}
}
\concept{StudentSubmission functions}
