% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{SubmissionHistory}
\alias{SubmissionHistory}
\title{SubmissionHistory Object}
\usage{
SubmissionHistory(stateHistory = NULL, gradeHistory = NULL)
}
\arguments{
\item{stateHistory}{The state history information of the submission, if present}

\item{gradeHistory}{The grade history information of the submission, if present}
}
\value{
SubmissionHistory object
}
\description{
SubmissionHistory Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The history of the submission. This currently includes state and gradehistories.
}
\concept{SubmissionHistory functions}
