% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Teacher}
\alias{Teacher}
\title{Teacher Object}
\usage{
Teacher(profile = NULL, userId = NULL, courseId = NULL)
}
\arguments{
\item{profile}{Global user information for the teacher}

\item{userId}{Identifier of the user}

\item{courseId}{Identifier of the course}
}
\value{
Teacher object
}
\description{
Teacher Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Teacher of a course.
}
\seealso{
Other Teacher functions: 
\code{\link{courses.teachers.create}()}
}
\concept{Teacher functions}
