% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{TimeOfDay}
\alias{TimeOfDay}
\title{TimeOfDay Object}
\usage{
TimeOfDay(seconds = NULL, minutes = NULL, hours = NULL, nanos = NULL)
}
\arguments{
\item{seconds}{Seconds of minutes of the time}

\item{minutes}{Minutes of hour of day}

\item{hours}{Hours of day in 24 hour format}

\item{nanos}{Fractions of seconds in nanoseconds}
}
\value{
TimeOfDay object
}
\description{
TimeOfDay Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a time of day. The date and time zone are either not significantor are specified elsewhere. An API may choose to allow leap seconds. Relatedtypes are google.type.Date and \code{google.protobuf.Timestamp}.
}
\concept{TimeOfDay functions}
