% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Topic}
\alias{Topic}
\title{Topic Object}
\usage{
Topic(courseId = NULL, updateTime = NULL, name = NULL, topicId = NULL)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{updateTime}{The time the topic was last updated by the system}

\item{name}{The name of the topic, generated by the user}

\item{topicId}{Unique identifier for the topic}
}
\value{
Topic object
}
\description{
Topic Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Topic created by a teacher for the course
}
\seealso{
Other Topic functions: 
\code{\link{courses.topics.create}()},
\code{\link{courses.topics.patch}()}
}
\concept{Topic functions}
