% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{userProfiles.guardianInvitations.patch}
\alias{userProfiles.guardianInvitations.patch}
\title{Modifies a guardian invitation.Currently, the only valid modification is to change the \code{state} from\code{PENDING} to \code{COMPLETE}. This has the effect of withdrawing the invitation.This method returns the following error codes:* \code{PERMISSION_DENIED} if the current user does not have permission to  manage guardians, if guardians are not enabled for the domain in question  or for other access errors.* \code{FAILED_PRECONDITION} if the guardian link is not in the \code{PENDING} state.* \code{INVALID_ARGUMENT} if the format of the student ID provided  cannot be recognized (it is not an email address, nor a \code{user_id} from  this API), or if the passed \code{GuardianInvitation} has a \code{state} other than  \code{COMPLETE}, or if it modifies fields other than \code{state}.* \code{NOT_FOUND} if the student ID provided is a valid student ID, but  Classroom has no record of that student, or if the \code{id} field does not  refer to a guardian invitation known to Classroom.}
\usage{
userProfiles.guardianInvitations.patch(
  GuardianInvitation,
  studentId,
  invitationId,
  updateMask = NULL
)
}
\arguments{
\item{GuardianInvitation}{The \link{GuardianInvitation} object to pass to this method}

\item{studentId}{The ID of the student whose guardian invitation is to be modified}

\item{invitationId}{The \code{id} field of the \code{GuardianInvitation} to be modified}

\item{updateMask}{Mask that identifies which fields on the course to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.guardianlinks.students
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.guardianlinks.students)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}

Other GuardianInvitation functions: 
\code{\link{GuardianInvitation}()},
\code{\link{userProfiles.guardianInvitations.create}()}
}
\concept{GuardianInvitation functions}
