% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{userProfiles.guardians.list}
\alias{userProfiles.guardians.list}
\title{Returns a list of guardians that the requesting user is permitted toview, restricted to those that match the request.To list guardians for any student that the requesting user may viewguardians for, use the literal character \code{-} for the student ID.This method returns the following error codes:* \code{PERMISSION_DENIED} if a \code{student_id} is specified, and the requesting  user is not permitted to view guardian information for that student, if  \code{'-'} is specified as the \code{student_id} and the user is not a domain  administrator, if guardians are not enabled for the domain in question,  if the \code{invited_email_address} filter is set by a user who is not a  domain administrator, or for other access errors.* \code{INVALID_ARGUMENT} if a \code{student_id} is specified, but its format cannot  be recognized (it is not an email address, nor a \code{student_id} from the  API, nor the literal string \code{me}). May also be returned if an invalid  \code{page_token} is provided.* \code{NOT_FOUND} if a \code{student_id} is specified, and its format can be  recognized, but Classroom has no record of that student.}
\usage{
userProfiles.guardians.list(
  studentId,
  pageToken = NULL,
  invitedEmailAddress = NULL,
  pageSize = NULL
)
}
\arguments{
\item{studentId}{Filter results by the student who the guardian is linked to}

\item{pageToken}{nextPageToken}

\item{invitedEmailAddress}{Filter results by the email address that the original invitation was sent}

\item{pageSize}{Maximum number of items to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/classroom.guardianlinks.me.readonly
\item https://www.googleapis.com/auth/classroom.guardianlinks.students
\item https://www.googleapis.com/auth/classroom.guardianlinks.students.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.guardianlinks.me.readonly, https://www.googleapis.com/auth/classroom.guardianlinks.students, https://www.googleapis.com/auth/classroom.guardianlinks.students.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
